# backend/src/openapi-generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MicroPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MicroPay — instant micropayments on the Stellar network.\n\n## Rate Limiting\n\nAll endpoints are rate-limited. Two limiters apply:\n\n| Limiter | Window | Limit | Routes |\n|---------|--------|-------|--------|\n| Global | 15 minutes | 100 req/IP | All routes |\n| Strict | 1 minute | 20 req/IP | `/api/turrets/*`, `/api/webhooks/*` |\n\nEvery response includes the following headers so clients can implement back-off:\n\n| Header | Description |\n|--------|-------------|\n| `RateLimit-Limit` | Maximum requests allowed in the current window |\n| `RateLimit-Remaining` | Requests remaining before the limit is reached |\n| `RateLimit-Reset` | Seconds until the window resets |\n\nWhen the limit is exceeded the server returns **HTTP 429** with `{ \"error\": \"Too many requests, please try again later.\" }`. Clients should read `RateLimit-Remaining` on each response and add exponential back-off when the value approaches 0.",
    "contact": {
      "name": "Stellar MicroPay",
      "url": "https://github.com/Emmy123222/Stellar-MicroPay"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/auth": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get SEP-0010 challenge transaction",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            },
            "description": "Stellar public key"
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge transaction XDR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": {
                      "type": "string"
                    },
                    "networkPassphrase": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid account parameter"
          }
        }
      },
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify signed challenge and get JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transaction": {
                    "type": "string",
                    "description": "Signed challenge XDR"
                  }
                },
                "required": [
                  "transaction"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid signature"
          }
        }
      }
    },
...
    "/api/tips": {
      "post": {
        "tags": [
          "Tips"
        ],
        "summary": "Record a new tip",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "string"
                  },
                  "memo": {
                    "type": "string"
                  },
                  "transactionHash": {
                    "type": "string"
                  }
                },
                "required": [
                  "from",
                  "to",
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tip recorded"
          },
          "400": {
            "description": "Invalid tip data"
          }
        }
      }
    },
    "/api/turrets": {
      "get": {
        "tags": [
          "Turrets"
        ],
        "summary": "List txFunction deployments",
        "description": "Returns all deployments. Filter by owner using `ownerPublicKey` query parameter.",
        "parameters": [
          {
            "name": "ownerPublicKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            },
            "description": "Filter deployments by owner Stellar public key"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "ownerPublicKey": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "dca",
                              "stop_loss",
                              "escrow_release"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "paused",
                              "completed"
                            ]
                          },
                          "config": {
                            "type": "object"
                          },
                          "deploymentHash": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "nextRunAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastExecutedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastCheckedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastObservedPriceUsd": {
                            "type": "number",
                            "nullable": true
                          },
                          "lastError": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/api/webhooks": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Register a new webhook",
        "description": "Register a webhook to receive notifications when payments occur for a specific Stellar public key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "publicKey",
                  "url",
                  "secret"
                ],
                "properties": {
                  "publicKey": {
                    "type": "string",
                    "pattern": "^G[A-Z0-9]{55}$",
                    "description": "Stellar public key to monitor"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "HTTPS URL to receive webhook notifications"
                  },
                  "secret": {
                    "type": "string",
                    "minLength": 32,
                    "description": "HMAC signing secret (min 32 chars)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "webhook": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "publicKey": {
                          "type": "string",
                          "description": "Stellar public key being monitored"
                        },
                        "url": {
                          "type": "string",
                          "description": "Destination URL for POST notifications"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/federation": {
      "get": {
        "tags": [
          "Federation"
        ],
        "summary": "SEP-0002 federation endpoint",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Federation query. Use user*domain for type=name or a public key for type=id."
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "id"
              ]
            },
            "description": "Query type"
          }
        ],
        "responses": {
          "200": {
            "description": "Federation record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stellar_address": {
                      "type": "string",
                      "example": "alice*stellarmicropay.io"
                    },
                    "account_id": {
                      "type": "string",
                      "example": "GABC...XYZ"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid federation query"
          },
          "404": {
            "description": "Federation record not found"
          }
        }
      }
    }
  }
}